Namespace Postgres Advisory locks to avoid collisions (#75)
Advisory locks are referenced by number, so we must avoid collisions
eg. lock '0' being used for two different purposes within the same
application.

Instead of referring to locks as a single long, we can refer to them
as two ints and treat the first as a namespace, for which we use a
random constant.

+++ src/main/java/uk/gov/hmcts/reform/sendletter/tasks/Task.java
@@ -4,7 +4,7 @@ public enum Task {
     UploadLetters,
     MarkLettersPosted;
 
-    long getLockId() {
+    int getLockId() {
         return ordinal();
     }
 }
